final connections

+++ lib/Data/styles/customexpansionpanel.dart
@@ -571,7 +571,7 @@ class _ExpansionPanelListState extends State<ExpansionPanelList> {
     return MergeableMaterial(
       hasDividers: true,
       dividerColor: widget.dividerColor,
-      elevation: widget.elevation,
+      elevation: widget.elevation.toDouble(),
       children: items,
     );
   }
